44 - Flash attributes / Flash messenger

+++ src/main/java/com/bolsadeideas/springboot/app/controllers/ClienteController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.bind.support.SessionStatus;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.bolsadeideas.springboot.app.models.entity.Cliente;
 import com.bolsadeideas.springboot.app.models.service.IClienteService;
@@ -41,12 +42,17 @@ public class ClienteController {
 	}
 	
 	@RequestMapping(value="/form/{id}")
-	public String editar(@PathVariable(value="id") Long id ,Map<String,Object> model) {
+	public String editar(@PathVariable(value="id") Long id ,Map<String,Object> model, RedirectAttributes flash) {
 		Cliente cliente = null;
 		if (id>0) {
 			cliente = clienteService.findOne(id);
+			if (cliente == null) {
+				flash.addFlashAttribute("error", "El ID del cliente no existe en la BBDD");
+				return "redirect:/listar";				
+			}
 		} else {
-			return "redirect:listar";
+			flash.addFlashAttribute("error", "El ID del cliente no puede ser cero");
+			return "redirect:/listar";
 		}
 		model.put("cliente", cliente);
 		model.put("titulo", "Editar cliente");
@@ -54,21 +60,33 @@ public class ClienteController {
 	}
 	
 	@RequestMapping(value="/form", method=RequestMethod.POST) 
-	public String guardar(@Valid Cliente cliente, BindingResult result, Map<String,Object> model, SessionStatus status) {
+	public String guardar(@Valid Cliente cliente, BindingResult result, Map<String,Object> model, 
+			RedirectAttributes flash, SessionStatus status) {
 		if(result.hasErrors()) {
 			model.put("titulo", "Formulario de cliente");
 			return "form";
 			}
+		String msgFlash = (cliente.getId() != null) ? "Cliente editado con exito" : "Cliente creado con exito" ;
 		clienteService.save(cliente);
 		status.setComplete();
+		flash.addFlashAttribute("success", msgFlash);
 		return "redirect:listar";
 	}
 	
 	@RequestMapping(value="/eliminar/{id}")
-	public String eliminar(@PathVariable(value="id") Long id) {
+	public String eliminar(@PathVariable(value="id") Long id, RedirectAttributes flash) {
 		if (id>0) {
-			clienteService.delete(id);
-		} 
+			Cliente cliente = clienteService.findOne(id);
+			if (cliente == null) {
+				flash.addFlashAttribute("error", "El ID del cliente no existe en la BBDD");
+				
+			} else {
+				clienteService.delete(id);
+				flash.addFlashAttribute("success", "Cliente eliminado con exito");
+			}
+		} else {
+			flash.addFlashAttribute("error", "El ID del cliente no puede ser cero");
+		}
 		return "redirect:/listar";
 	}
 }
